backend/qercas_project/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8
requests>=2.31.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0
numpy>=1.24.0
shap>=0.48.0
networkx>=3.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
qiskit>=0.45.0
qiskit-aer>=0.12.0
pycryptodome>=3.19.0
faker>=19.0.0
